requirements.txt
sentence-transformers==2.2.2
diffusers==0.25.1
transformers==4.35.2
accelerate==0.22.0
peft==0.7.0
bitsandbytes==0.46.0
huggingface-hub==0.21.4
torch==2.6.0
pyopenssl
matplotlib
numpy
Pillow
tqdm
    